refactor(movie-app): move App.js logic to new component MovieApp.js
