+++ movie-app/src/components/index.js
@@ -1,0 +1,19 @@
+import Carousel from "./Carousel";
+import Header from "./Header";
+import Modal from "./Modal";
+import MovieApp from "./MovieApp";
+import MovieCard from "./MovieCard";
+import MovieDetails from "./MovieDetails";
+import MovieList from "./MovieList";
+import PaginationBar from "./PaginationBar";
+
+export {
+    Carousel,
+    Header,
+    Modal,
+    MovieApp,
+    MovieCard,
+    MovieDetails,
+    MovieList,
+    PaginationBar
+}
